Refactors decomp.semantics.uds and corresponding tests into multiple submodules and moves to pytest as testing framework
