+++ decomp/__init__.py
@@ -8,4 +8,6 @@ basicConfig(filename=os.path.join(DATA_DIR, 'build.log'),
             filemode='w',
             level=DEBUG)
 
-from .semantics.uds import UDSCorpus, NormalizedUDSDataset, RawUDSDataset
+from .semantics.uds import UDSCorpus
+from .semantics.uds import NormalizedUDSAnnotation
+from .semantics.uds import RawUDSAnnotation
