+++ tests/data/normalized_edge_sentence_annotation.json
@@ -1,0 +1,1 @@
+{"metadata": {"protoroles": {"awareness": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "change_of_location": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "change_of_possession": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "change_of_state": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "change_of_state_continuous": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "existed_after": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "existed_before": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "existed_during": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "instigation": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "location": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "manner": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "partitive": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "purpose": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "sentient": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "time": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "volition": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "was_for_benefit": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}, "was_used": {"value": {"datatype": "float"}, "confidence": {"datatype": "float"}}}}, "data": {"tree1": {"tree1-semantics-pred-11%%tree1-semantics-arg-13": {"protoroles": {"instigation": {"confidence": 1.0, "value": -0.0}, "change_of_possession": {"confidence": 1.0, "value": -0.0}, "existed_before": {"confidence": 0.6796, "value": 0.0111}, "was_for_benefit": {"confidence": 1.0, "value": -0.0}, "change_of_state_continuous": {"confidence": 0.1675, "value": 0.0032}, "change_of_state": {"confidence": 0.1675, "value": 0.0032}, "volition": {"confidence": 1.0, "value": -0.0}, "change_of_location": {"confidence": 1.0, "value": -0.0}, "partitive": {"confidence": 0.564, "value": -0.0941}, "existed_during": {"confidence": 1.0, "value": 1.3421}, "existed_after": {"confidence": 0.6796, "value": 0.0111}, "awareness": {"confidence": 1.0, "value": -0.0}, "sentient": {"confidence": 1.0, "value": -0.9348}, "was_used": {"confidence": 0.564, "value": -0.0}}}, "tree1-semantics-pred-7%%tree1-semantics-arg-3": {"protoroles": {"instigation": {"confidence": 1.0, "value": 1.3557}, "change_of_possession": {"confidence": 0.7724, "value": -0.0}, "existed_before": {"confidence": 1.0, "value": 1.3527}, "was_for_benefit": {"confidence": 0.1976, "value": -0.0504}, "change_of_state_continuous": {"confidence": 1.0, "value": -0.0}, "change_of_state": {"confidence": 0.2067, "value": -0.0548}, "volition": {"confidence": 1.0, "value": 1.3545}, "change_of_location": {"confidence": 0.272, "value": -0.0922}, "partitive": {"confidence": 0.1148, "value": -0.0018}, "existed_during": {"confidence": 1.0, "value": 1.3557}, "existed_after": {"confidence": 1.0, "value": 1.3527}, "awareness": {"confidence": 1.0, "value": 1.3526}, "sentient": {"confidence": 1.0, "value": 1.354}, "was_used": {"confidence": 0.4373, "value": -0.0207}}}, "tree1-semantics-pred-11%%tree1-semantics-arg-9": {"protoroles": {"instigation": {"confidence": 1.0, "value": -1.5074}, "change_of_possession": {"confidence": 1.0, "value": -0.3909}, "existed_before": {"confidence": 1.0, "value": 1.3954}, "was_for_benefit": {"confidence": 0.3418, "value": 0.0008}, "change_of_state_continuous": {"confidence": 0.0791, "value": -0.0351}, "change_of_state": {"confidence": 0.3333, "value": -0.0085}, "volition": {"confidence": 1.0, "value": -0.3909}, "change_of_location": {"confidence": 0.1395, "value": -0.0549}, "partitive": {"confidence": 0.0791, "value": -0.1354}, "existed_during": {"confidence": 1.0, "value": 1.3959}, "existed_after": {"confidence": 0.6567, "value": 0.124}, "awareness": {"confidence": 0.1395, "value": -0.0549}, "sentient": {"confidence": 1.0, "value": -1.508}, "was_used": {"confidence": 0.3333, "value": -0.0085}}}}}}
