+++ tests/data/raw_edge_sentence_annotators.txt
@@ -1,0 +1,46 @@
+protoroles-annotator-0
+protoroles-annotator-1
+protoroles-annotator-10
+protoroles-annotator-11
+protoroles-annotator-12
+protoroles-annotator-13
+protoroles-annotator-14
+protoroles-annotator-15
+protoroles-annotator-16
+protoroles-annotator-17
+protoroles-annotator-18
+protoroles-annotator-19
+protoroles-annotator-2
+protoroles-annotator-20
+protoroles-annotator-21
+protoroles-annotator-22
+protoroles-annotator-23
+protoroles-annotator-24
+protoroles-annotator-25
+protoroles-annotator-26
+protoroles-annotator-27
+protoroles-annotator-28
+protoroles-annotator-29
+protoroles-annotator-3
+protoroles-annotator-30
+protoroles-annotator-31
+protoroles-annotator-32
+protoroles-annotator-33
+protoroles-annotator-34
+protoroles-annotator-35
+protoroles-annotator-36
+protoroles-annotator-37
+protoroles-annotator-38
+protoroles-annotator-39
+protoroles-annotator-4
+protoroles-annotator-40
+protoroles-annotator-41
+protoroles-annotator-42
+protoroles-annotator-43
+protoroles-annotator-44
+protoroles-annotator-45
+protoroles-annotator-5
+protoroles-annotator-6
+protoroles-annotator-7
+protoroles-annotator-8
+protoroles-annotator-9
