+++ tests/data/rawtree.conllu
@@ -1,0 +1,29 @@
+1	The	the	DET	DT	Definite=Def|PronType=Art	3	det	_	_
+2	police	police	NOUN	NN	Number=Sing	3	compound	_	_
+3	commander	commander	NOUN	NN	Number=Sing	7	nsubj	_	_
+4	of	of	ADP	IN	_	6	case	_	_
+5	Ninevah	Ninevah	PROPN	NNP	Number=Sing	6	compound	_	_
+6	Province	Province	PROPN	NNP	Number=Sing	3	nmod	_	_
+7	announced	announce	VERB	VBD	Mood=Ind|Tense=Past|VerbForm=Fin	0	root	_	_
+8	that	that	SCONJ	IN	_	11	mark	_	_
+9	bombings	bombing	NOUN	NNS	Number=Plur	11	nsubj	_	_
+10	had	have	AUX	VBD	Mood=Ind|Tense=Past|VerbForm=Fin	11	aux	_	_
+11	declined	decline	VERB	VBN	Tense=Past|VerbForm=Part	7	ccomp	_	_
+12	80	80	NUM	CD	NumType=Card	13	nummod	_	_
+13	percent	percent	NOUN	NN	Number=Sing	11	dobj	_	_
+14	in	in	ADP	IN	_	15	case	_	_
+15	Mosul	Mosul	PROPN	NNP	Number=Sing	11	nmod	_	SpaceAfter=No
+16	,	,	PUNCT	,	_	11	punct	_	_
+17	whereas	whereas	SCONJ	IN	_	20	mark	_	_
+18	there	there	PRON	EX	_	20	expl	_	_
+19	had	have	AUX	VBD	Mood=Ind|Tense=Past|VerbForm=Fin	20	aux	_	_
+20	been	be	VERB	VBN	Tense=Past|VerbForm=Part	11	advcl	_	_
+21	a	a	DET	DT	Definite=Ind|PronType=Art	23	det	_	_
+22	big	big	ADJ	JJ	Degree=Pos	23	amod	_	_
+23	jump	jump	NOUN	NN	Number=Sing	20	nsubj	_	_
+24	in	in	ADP	IN	_	26	case	_	_
+25	the	the	DET	DT	Definite=Def|PronType=Art	26	det	_	_
+26	number	number	NOUN	NN	Number=Sing	23	nmod	_	_
+27	of	of	ADP	IN	_	28	case	_	_
+28	kidnappings	kidnapping	NOUN	NNS	Number=Plur	26	nmod	_	SpaceAfter=No
+29	.	.	PUNCT	.	_	7	punct	_	_
